websearch/requirement.txt
streamlit
yfinance==1.0
tavily-python
python-dotenv
langchain==1.2.0
langchain-core==1.2.5
langchain-groq
langchain-community
requests